requirements.txt
Django
djangorestframework
psycopg2-binary
django-environ
django-allauth
dj-rest-auth
django-fernet-fields @ git+https://github.com/badri/django-fernet-fields.git@django-4.2
